requirements.txt
Django~=5.1.3
django-redis~=5.4.0
djangorestframework~=3.15.2
djangorestframework-simplejwt~=5.3.1
drf-spectacular~=0.27.2
drf-yasg~=1.21.8
python-decouple~=3.8
pytz~=2024.2
redis~=5.2.0
